# docs/swagger.yaml
basePath: /
definitions:
...
  dto.RegisterWebhookRequest:
    properties:
      events:
        example:
        - Message
        - Receipt
        - Connected
        items:
          type: string
        type: array
      secret:
        example: webhook_secret_key
        type: string
      url:
        example: https://webhook.example.com/meow
        type: string
    required:
    - events
    - url
    type: object
...
  dto.StandardWebhookCreateResponse:
    properties:
      data:
        $ref: '#/definitions/dto.StandardWebhookData'
      message:
        example: Webhook created successfully
        type: string
      status:
        example: 201
        type: integer
    type: object
  dto.StandardWebhookData:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      events:
        example:
        - Message
        - Receipt
        - Connected
        items:
          type: string
        type: array
      sessionID:
        example: default
        type: string
      status:
        example: active
        type: string
      url:
        example: https://webhook.example.com/meow
        type: string
    type: object
  dto.StandardWebhookResponse:
    properties:
      data:
        $ref: '#/definitions/dto.StandardWebhookData'
      message:
        example: Webhook retrieved successfully
        type: string
      status:
        example: 200
        type: integer
    type: object
...
  dto.SupportedEventsData:
    properties:
      count:
        example: 65
        type: integer
      events:
        example:
        - '["Message"'
        - ' "Receipt"'
        - ' "Connected"]'
        items:
          type: string
        type: array
    type: object
  dto.SupportedEventsResponse:
    properties:
      data:
        $ref: '#/definitions/dto.SupportedEventsData'
      message:
        example: Supported events retrieved successfully
        type: string
      status:
        example: 200
        type: integer
    type: object
  dto.TestWebhookResult:
    properties:
      error:
        example: ""
        type: string
      response_code:
        example: 200
        type: integer
      response_time_ms:
        example: 150
        type: integer
      test_result:
        example: success
        type: string
      webhook_id:
        example: webhook_123456789
        type: string
    type: object
...
  dto.WebhookErrorResponse:
    properties:
      code:
        example: INVALID_WEBHOOK_URL
        type: string
      details:
        example: Webhook URL must be a valid HTTPS URL
        type: string
      message:
        example: Invalid webhook URL format
        type: string
    type: object
  dto.WebhookInfo:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      events:
        example:
        - Message
        - Receipt
        - Connected
        items:
          type: string
        type: array
      secret:
        example: webhook_secret_key
        type: string
      session_id:
        example: default
        type: string
      status:
        example: active
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      url:
        example: https://webhook.example.com/meow
        type: string
      webhook_id:
        example: webhook_123456789
        type: string
    type: object
  dto.WebhookResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.WebhookResponseData'
      error:
        $ref: '#/definitions/dto.WebhookErrorResponse'
      success:
        type: boolean
    type: object
  dto.WebhookResponseData:
    properties:
      action:
        example: register
        type: string
      status:
        example: success
        type: string
      test_result:
        $ref: '#/definitions/dto.TestWebhookResult'
      timestamp:
        example: "2023-01-01T00:00:00Z"
        type: string
      webhook:
        $ref: '#/definitions/dto.WebhookInfo'
      webhook_id:
        example: webhook_123456789
        type: string
      webhooks:
        items:
          $ref: '#/definitions/dto.WebhookInfo'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@meow.com
    name: meow API Support
    url: https://github.com/your-username/meow
  description: A meow API server built with Go, inspired by meow
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: meow meow API
  version: "1.0"
paths:
...
  /session/{sessionId}/webhook:
    get:
      consumes:
      - application/json
      description: Get information about registered webhooks for a session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StandardWebhookResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.WebhookResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.WebhookResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.WebhookResponse'
      security:
      - ApiKeyAuth: []
      summary: Get webhook information
      tags:
      - Webhooks
    post:
      consumes:
      - application/json
      description: Set a webhook URL to receive meow events
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Set webhook request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterWebhookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.StandardWebhookCreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.WebhookResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.WebhookResponse'
      security:
      - ApiKeyAuth: []
      summary: Set webhook
      tags:
      - Webhooks
  /session/{sessionId}/webhooks/events:
    get:
      consumes:
      - application/json
      description: Get list of all supported webhook event types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SupportedEventsResponse'
      security:
      - ApiKeyAuth: []
      summary: List supported events
      tags:
      - Webhooks
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: 'API Key authentication. Simply provide your API key directly: "YOUR_API_KEY".
      The system automatically detects if it''s a Global API Key (can access all sessions
      and session management) or a Session-specific API Key (can only access the specific
      session it belongs to).'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...